style: move analyses and init-tsconfig
