+++ src/lib/ng-package/entry-point/analyse-sources.di.ts
@@ -1,6 +1,6 @@
 import { InjectionToken } from 'injection-js';
-import { Transform } from '../graph/transform';
-import { TransformProvider, provideTransform } from '../graph/transform.di';
+import { Transform } from '../../graph/transform';
+import { TransformProvider, provideTransform } from '../../graph/transform.di';
 import { analyseSourcesTransform } from './analyse-sources.transform';
 
 export const ANALYSE_SOURCES_TOKEN = new InjectionToken<Transform>(`ng.v5.analyseSourcesTransform`);
